refactor(android): address code review feedback for chart interactions
- add try/catch with rollback in SupplementsWidget (Issue 1)
- remove redundant re-fetch after optimistic update (Issue 2)
- hoist Paint allocations outside Canvas draw scope in WeightTrendChart and MacroRadarChart (Issue 3)
- use real measured tooltip size via onSizeChanged in ChartTooltip (Issue 4)
- use string resources for hardcoded chart strings (Issue 6)
- extract SimpleLineChart and SimplePieChart to separate component files (Issue 7)
- move MealColors to SimplePieChart.kt as public val
